+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Revision history for swankybar
 
+## 0.1.9.1 -- 2023-08-07
+
+* Added `swankybar-ttc` for fetching the next TTC ETAs
+
+## 0.1.9.0 -- 2023-08-06
+
+* Added `swankybar-cpugov` to display the current CPU governor
+
 ## 0.1.0.0 -- 2023-07-28
 
 * First version. Released on an unsuspecting world!